pyproject.toml
[project]
name = "gtsrb-sign-recognition"
version = "0.1.0"
description = "Traffic sign recognition with GTSRB"
authors = [
    {name = "MateiSapunaru",email = "sapunarumatei@gmail.com"}
]
requires-python = ">=3.11,<3.14.1"
dependencies = [
    "streamlit (>=1.56.0,<2.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "torch (>=2.11.0,<3.0.0)",
    "torchvision (>=0.26.0,<0.27.0)",
    "torchaudio (>=2.11.0,<3.0.0)",
    "grad-cam (>=1.5.5,<2.0.0)",
    "streamlit-webrtc (>=0.64.5,<0.65.0)",
    "albumentations (>=2.0.8,<3.0.0)",
    "timm (>=1.0.26,<2.0.0)",
    "plotly (>=6.7.0,<7.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "av (>=14,<17)",
    "requests (>=2.33.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "notebook (>=7.5.5,<8.0.0)"
]

[[tool.poetry.source]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-cu128"}
torchvision = {source = "pytorch-cu128"}
torchaudio = {source = "pytorch-cu128"}
